+++ include/neuralnetwork/layers/fflayer.cpp
@@ -1052,16 +1052,7 @@ void FFLayer::cache_recurrent_weights()
     return;
   }
   _w_values_T.resize(n_this * n_prev);
-  const double* w_ptr = w_vals.data();
-  double* wt_ptr = _w_values_T.data();
-  for (size_t i = 0; i < n_prev; ++i)
-  {
-    const double* src_row = w_ptr + i * n_this;
-    for (size_t j = 0; j < n_this; ++j)
-    {
-      wt_ptr[j * n_prev + i] = src_row[j];
-    }
-  }
+  simd::transpose(w_vals.data(), _w_values_T.data(), n_prev, n_this);
 }
 
 void FFLayer::run_gemm_backward_fast(